+++ aggregation/src/main/java/com/tgdating/aggregation/dto/request/RequestProfileListGetDto.java
@@ -9,8 +9,8 @@ public class RequestProfileListGetDto {
     private String sessionId;
     private String searchGender;
     private String lookingFor;
-    private Byte ageFrom;
-    private Byte ageTo;
+    private Integer ageFrom;
+    private Integer ageTo;
     private Double distance;
     private Double latitude;
     private Double longitude;
