+++ aggregation/src/main/java/com/tgdating/aggregation/model/ProfileFilterEntity.java
@@ -12,8 +12,8 @@ public class ProfileFilterEntity {
     private String sessionId;
     private String searchGender;
     private String lookingFor;
-    private Byte ageFrom;
-    private Byte ageTo;
+    private Integer ageFrom;
+    private Integer ageTo;
     private Double distance;
     private Integer page;
     private Integer size;
